Move both async generatrs into common iterators module
